+++ src/main/java/de/retest/web/RecheckSeleniumAdapter.java
@@ -53,12 +53,17 @@ public class RecheckSeleniumAdapter implements RecheckAdapter {
 
 	public RecheckSeleniumAdapter( final RecheckOptions options ) {
 		retestIdProvider = options.getRetestIdProvider();
+		screenshotProvider = getScreenshotProvider( options );
+	}
+
+	private ScreenshotProvider getScreenshotProvider( final RecheckOptions options ) {
 		if ( options instanceof RecheckWebOptions ) {
-			final RecheckWebOptions webOptions = (RecheckWebOptions) options;
-			screenshotProvider = webOptions.getScreenshotProvider();
-		} else {
-			screenshotProvider = RecheckWebProperties.getInstance().screenshotProvider();
+			final ScreenshotProvider fromOptions = ((RecheckWebOptions) options).getScreenshotProvider();
+			if ( fromOptions != null ) {
+				return fromOptions;
+			}
 		}
+		return RecheckWebProperties.getInstance().screenshotProvider();
 	}
 
 	public RecheckSeleniumAdapter() {
@@ -67,7 +72,6 @@ public class RecheckSeleniumAdapter implements RecheckAdapter {
 
 	@Override
 	public RecheckAdapter initialize( final RecheckOptions opts ) {
-		RecheckWebProperties.init();
 		return new RecheckSeleniumAdapter( opts );
 	}
 
@@ -117,8 +121,8 @@ public class RecheckSeleniumAdapter implements RecheckAdapter {
 	private Object unwrapImplicitDriver( final Object toVerify ) {
 		if ( toVerify instanceof AutocheckingRecheckDriver ) {
 			throw new UnsupportedOperationException( String.format(
-					"The '%s' does implicit checking after each action, " // 
-							+ "therefore no explicit check with 'Recheck#check' is needed. " // 
+					"The '%s' does implicit checking after each action, " //
+							+ "therefore no explicit check with 'Recheck#check' is needed. " //
 							+ "Please remove the explicit check. " //
 							+ "For more information, please have a look at https://docs.retest.de/recheck-web/introduction/usage/.",
 					toVerify.getClass().getSimpleName() ) );