+++ src/main/java/de/retest/web/RecheckWebProperties.java
@@ -4,10 +4,8 @@ import org.aeonbits.owner.Config.LoadPolicy;
 import org.aeonbits.owner.Config.LoadType;
 import org.aeonbits.owner.Config.Sources;
 import org.aeonbits.owner.ConfigCache;
-import org.aeonbits.owner.ConfigFactory;
 import org.aeonbits.owner.Reloadable;
 
-import de.retest.recheck.configuration.ProjectRootFinderUtil;
 import de.retest.web.screenshot.ScreenshotProvider;
 import de.retest.web.screenshot.ScreenshotProviders.ScreenshotProviderConverter;
 
@@ -19,14 +17,8 @@ public interface RecheckWebProperties extends Reloadable {
 	 * Basic usage.
 	 */
 
-	static void init() {
-		ProjectRootFinderUtil.getProjectRoot().ifPresent(
-				projectRoot -> ConfigFactory.setProperty( "projectroot", projectRoot.toAbsolutePath().toString() ) );
-	}
-
 	static RecheckWebProperties getInstance() {
-		final RecheckWebProperties instance =
-				ConfigCache.getOrCreate( RecheckWebProperties.class, System.getProperties() );
+		final RecheckWebProperties instance = ConfigCache.getOrCreate( RecheckWebProperties.class );
 		instance.reload();
 		return instance;
 	}