+++ CHANGELOG.md
@@ -25,12 +25,13 @@ Table of Contents
 
 ### Breaking Changes
 
-* Add `shoot(WebDriver, WebElement)` in interface `ScreenshotProvider` to create screenshots for `WebElement`
+* Add `shoot(WebDriver, WebElement)` in interface `ScreenshotProvider` to create screenshots for `WebElement`.
 
 ### Bug Fixes
 
 * Fix that cascading frames inside of frames are not analyzed. Now depth doesn't matter...
-* Fix that `NoScreenshot` or `RecheckWebOptions#disableScreenshots` is not applied to web elements.
+* Fix that `NoScreenshot` or `RecheckWebOptions#disableScreenshots()` is not applied to web elements.
+* Fix that `de.retest.recheck.web.screenshot.provider` from `recheck.ignore` is ignored when `RecheckWebOptions` are used.
 
 ### New Features